+++ apps/client/src/pages/home/components/performance-carousel-section.css.ts
@@ -1,0 +1,9 @@
+import { style } from '@vanilla-extract/css';
+
+import { themeVars } from '@confeti/design-system/styles';
+
+export const performanceBannerContainer = style({
+  width: '100%',
+  paddingTop: '0.1rem',
+  background: themeVars.color.confeti_lime_grad,
+});
