+++ apps/client/src/pages/home/components/suggest-music-section.tsx
@@ -1,0 +1,16 @@
+import { Box } from '@confeti/design-system';
+
+const SuggestMusicSection = () => {
+  return (
+    <Box
+      title="미리 음악을 한 번 들어볼까요?"
+      titleSize="lg"
+      subtitle="♥ 인천 펜타포트 락 페스티벌"
+    >
+      {/* TODO: 음악 리스트 컴포넌트 추가 */}
+      <p>음악 리스트</p>
+    </Box>
+  );
+};
+
+export default SuggestMusicSection;
