+++ apps/client/src/pages/home/components/suggest-performance-section.tsx
@@ -1,0 +1,15 @@
+import { Box } from '@confeti/design-system';
+
+const SuggestPerformanceSection = () => {
+  return (
+    <Box
+      title="이런 공연은 어떠세요?"
+      titleSize="lg"
+      subtitle="♥ confeti's pick!"
+    >
+      <p>test</p>
+    </Box>
+  );
+};
+
+export default SuggestPerformanceSection;
