+++ apps/client/src/pages/home/page/home.tsx
@@ -1,52 +1,17 @@
 import { useEffect } from 'react';
 import { useSocialLoginMutation } from '@pages/login/hooks/use-social-login-mutation';
-import { useUserProfile } from '@pages/my/hooks/use-user-info';
 
-import {
-  Footer,
-  PerformanceCarousel,
-  TicketingCard,
-} from '@confeti/design-system';
+import { Footer, Spacing } from '@confeti/design-system';
 import NavigationTabs from '@shared/components/navigation-tabs';
 import { CONFIG } from '@shared/constants/api';
-import { useNavigateToDetail } from '@shared/hooks/use-navigate-to-detail';
-import { formatDate } from '@shared/utils/format-date';
 
+import PerformanceCarouselSection from '../components/performance-carousel-section';
+import SuggestMusicSection from '../components/suggest-music-section';
+import SuggestPerformanceSection from '../components/suggest-performance-section';
+import TicketingSection from '../components/ticketing-section';
 import { TAB_MENU } from '../constants/menu';
-import { useLatestPerformances } from '../hooks/use-latest-performances';
-import { useTicketing } from '../hooks/use-ticketing';
-
-import * as styles from './home.css';
-
-import ImgDday01 from '/images/img_dday01.svg';
-import ImgDday02 from '/images/img_dday02.svg';
-import ImgDday03 from '/images/img_dday03.svg';
-import ImgDday04 from '/images/img_dday04.svg';
-import ImgDday05 from '/images/img_dday05.svg';
 
 const Home = () => {
-  const { performances } = useTicketing();
-  const { latestPerformances } = useLatestPerformances();
-  const displayPerformances =
-    latestPerformances.length > 7
-      ? latestPerformances.slice(0, 7)
-      : latestPerformances;
-
-  const navigateToDetail = useNavigateToDetail();
-
-  const formattedPerformData = displayPerformances.map((performance) => ({
-    ...performance,
-    performanceAt: formatDate(performance.performanceAt),
-  }));
-  const DdayList = performances?.map((performance) => ({
-    ...performance,
-    reserveAt: formatDate(performance.reserveAt, 'Dday'),
-  }));
-  const { data: profileData } = useUserProfile();
-
-  const initialSlideIndex = Math.floor(formattedPerformData.length / 2);
-  const imageUrls = [ImgDday01, ImgDday02, ImgDday03, ImgDday04, ImgDday05];
-
   const { mutate: login } = useSocialLoginMutation();
   const params = new URLSearchParams(window.location.search);
   const code = params.get('code');
@@ -65,63 +30,17 @@ const Home = () => {
     <>
       <NavigationTabs defaultActiveTab={TAB_MENU.HOME} />
 
-      <div className={styles.background}>
-        <section className={styles.performanceBannerContainer}>
-          <PerformanceCarousel
-            performData={formattedPerformData}
-            initialSlideIndex={initialSlideIndex}
-          >
-            <PerformanceCarousel.ImageSlider>
-              <PerformanceCarousel.Badge text="선호하는 아티스트" />
-              <PerformanceCarousel.Info />
-            </PerformanceCarousel.ImageSlider>
-          </PerformanceCarousel>
-        </section>
+      <PerformanceCarouselSection />
+      <Spacing size="xl" color="white" />
 
-        <section className={styles.ticketingBannerContainer}>
-          <p className={styles.ticketingBannerText}>
-            {profileData ? (
-              <>
-                <span className={styles.highlightedText}>
-                  {profileData.name}
-                </span>
-                님 <br />
-                예매가 다가오고 있어요!
-              </>
-            ) : (
-              <>
-                공연의 시작과 끝을 <br /> 콘페티와 함께해보세요!
-              </>
-            )}
-          </p>
-          <div className={styles.ticketingCardContainer}>
-            {performances?.map((performance, index) => (
-              <TicketingCard.Image
-                key={performance.index}
-                imageUrl={imageUrls[index]}
-                textContent={
-                  <>
-                    <TicketingCard.Dday
-                      reserveAt={DdayList[index]?.reserveAt}
-                    />
-                    <TicketingCard.SubTitle subtitle={performance.subtitle} />
-                  </>
-                }
-                performanceInfoContent={
-                  <TicketingCard.PerformanceInfo
-                    title={'공연 정보 확인하기'}
-                    typeId={performance.typeId}
-                    performanceType={performance.type}
-                    onClick={() =>
-                      navigateToDetail(performance.type, performance.typeId)
-                    }
-                  />
-                }
-              />
-            ))}
-          </div>
-        </section>
-      </div>
+      <TicketingSection />
+      <Spacing size="2xl" color="white" />
+
+      <SuggestPerformanceSection />
+      <Spacing size="lg" color="white" />
+
+      <SuggestMusicSection />
+
       <Footer />
     </>
   );
